Rewrite `Joiner.on(", ").join("a", "b")` to `String.join(", ", "a", "b")`
Adds a recipe to rewrite calls to Guava Joiner#join with String.join in compatible cases. A call can be re-written if the delimiter is a String, and the arguments are an array, a list or an iterable of items that can be cast to a CharSequence.

Addresses https://github.com/openrewrite/rewrite-migrate-java/issues/318