ClassCastException in OptionalNotEmptyToIsPresent
This should fix #417, but I haven't been able to verify locally.

Don't assume the super call to still return an instance of Statement,
but rather take the most generic type of the super declaration.

Also move the matcher declaration into the visitor, since its unused
code unless we enter the inner parts of the visitor method.

The additional test is green with and without the main code change, so
it does NOT reproduce the bug.

Finally rename the previously existing test example to match the actual
test intention.