Handle generics in PreferJavaUtilObjectsRequireNonNull
Before this patch, the PreferJavaUtilObjectsRequireNonNull recipe
would only migrate usage of exactly `checkNotNull(Object)` but not
any other (sub)types.

We now match invocations of checkNotNull with exactly one or two
arguments (both in refaster). Note that there is also a three
argument version that doesn't have a straight-forward replacement,
I skipped that for now and added a testcase.