Always add explicit type to the `Predicate.isEqual` method (#927)
* Always "cast" the `Predicate.isEqual` method

In some cases, the type can be inferred, but not always.
To follow the “Do no harm” principle, it’s safer to always "cast" the
`Predicate.isEqual` method explicitly.
Before this commit, certain cases resulted in compilation errors (see
the new test inlinedPredicatesEqualToWithStandardType, which reveals
this problem).

* Add the recipe NoGuavaMapsNewLinkedHashMap to `no-guava.yml`

It seems that the recipe NoGuavaMapsNewLinkedHashMap has never been
added in `no-guava.yml`.

* Don't use try/catch to handle null type

---------

Co-authored-by: Tim te Beek <tim@moderne.io>